requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
email-validator
